running experiments for different configurations and evaluating the respective models
running cbuild.sh or on windows win-build.sh also starts the experiments. after training the various models, they are evaluated and for each dataset a csv with the respective results is created.

TODO: currently our_bits and #features and #leaves are only stored as zeros in csv as mrf->printForest() is not called